Functions

Functions

Merlin.BiLSTM

Merlin.Conv1D

Merlin.LSTM

Merlin.Linear

Base.:*

Base.:+

Base.:-

Base.:/

Base.:^

Base.broadcast

Base.getindex

Base.max

Base.reshape

Base.tanh

Base.transpose

Merlin.concat

Merlin.crelu

Merlin.crossentropy

Merlin.dropout

Merlin.elu

Merlin.logsoftmax

Merlin.mse

Merlin.relu

Merlin.selu

Merlin.sigmoid

Merlin.softmax

Merlin.softmax_crossentropy

Activation

Merlin.crelu — Function.

crelu(x)

Concatenated Rectified Linear Unit. The output is twice the size of the input.

\[f(x) = (\max(0,x), \max(0,-x))\]

source

Merlin.elu — Function.

elu(x)

Exponential Linear Unit.

\[f(x) = \begin{cases} x & x > 0 \\ \alpha (e^{x}-1) & x\leq0 \end{cases}\]

source

Merlin.relu — Function.

relu(x)

Rectified Linear Unit.

\[f(x) = max(0, x)\]

source

Merlin.selu — Function.

selu(x)

Scaled Exponential Linear Unit.

\[f(x) = \lambda \begin{cases} x & x > 0 \\ \alpha e^{x}-\alpha & x\leq0 \end{cases}\]

References

Klambauer et al., Self-Normalizing Neural Networks, NIPS 2017

source

Merlin.sigmoid — Function.

sigmoid(x)

Sigmoid logistic function.

\[f(x) = (1 + \exp(-x))^{-1}\]

source

Base.tanh — Function.

tanh(x)

Hyperbolic tangent function.

source

Convolution

Merlin.Conv1D — Type.

Conv1D(T, ksize, insize, outsize, pad, stride; dilation=1, [init_w=Xavier()], [init_b=Zeros()])

1-dimensional convolution function.

x = Var(rand(Float32,10,5))
f = Conv1D(Float32, 5, 10, 3, 2, 1)
y = f(x)

source

Loss

Merlin.crossentropy — Function.

crossentropy(p, q)

Cross-entropy function between p and q.

\[f(x) = -\sum_{x} p(x) \log q(x)\]

p::Var: Var of Vector{Int} or Matrix{Float}. If p is Vector{Int} and p[i] == 0, returns 0.

q::Var: Var of Matrix{Float}

p = Var(rand(0:10,5))
q = softmax(Var(rand(Float32,10,5)))
y = crossentropy(p, q)

source

Merlin.mse — Function.

mse(x1, x2)

Mean Squared Error function between x1 and x2. The mean is calculated over the minibatch. Note that the error is not scaled by 1/2.

source

Merlin.softmax_crossentropy — Function.

softmax_crossentropy(p, x)

Cross-entropy function between p and $softmax(x)$.

\[f(x) = -\sum_{x} p(x) \log q(x)\]

where $q = softmax(x)$

p: Var of Vector{Int} or Matrix{Float}

q: Var of Matrix{Float}

p = Var(rand(0:10,5))
q = Var(rand(Float32,10,5))
y = softmax_crossentropy(p, x)

source

Math

Base.broadcast — Function.

.+(x1::Var, x2::Var)

source

.-(x1::Var, x2::Var)

source

.*(x1::Var, x2::Var)

source

Base.:+ — Function.

+(x1::Var, x2::Var)
+(a::Number, x::Var)
+(x::Var, a::Number)

source

Base.:- — Function.

-(x1, x2)

source

Base.:* — Function.

*(A::Var, B::Var)

source

Base.:/ — Function.

/(x1::Var, a)

source

Base.:^ — Function.

^(x::Var, a::Number)

source

Base.transpose — Function.

transpose(x)

source

Random

Merlin.dropout — Function.

dropout(x::Var, rate::Float64)

If config.train is true, drops elements randomly with probability $rate$ and scales the other elements by factor $1 / (1 - rate)$. Otherwise, it just returns x.

source

Recurrent

Merlin.BiLSTM — Type.

BiLSTM(x)

Bidirectional Long Short-Term Memory network.

👉 Example

x = Var(rand(Float32,10,5))

source

Merlin.LSTM — Type.

LSTM(::Type{T}, insize::Int, outsize::Int, [init_W=Uniform(0.001), init_U=Orthogonal()])

Long Short-Term Memory network.

T = Float32
f = LSTM(T, 100, 100)
h = f(x)

source

Reduction

max
max_batch

gemm
gemv
concat
getindex
Linear
logsoftmax
lookup
reshape
softmax
standardize
window1d