Optimizers

Optimizers

Merlin.AdaGrad

Merlin.Adam

Merlin.SGD

Optimizers provides a way to update the weights of Merlin.Var.

x = zerograd(rand(Float32,5,4))
opt = SGD(0.001)
opt(x)
println(x.grad)

Merlin.AdaGrad — Type.

AdaGrad

AdaGrad Optimizer.

References

Duchi t al., "Adaptive Subgradient Methods for Online Learning and Stochastic Optimization", JMLR 2011.

source

Merlin.Adam — Type.

Adam

Adam Optimizer

References

Kingma and Ba, "Adam: A Method for Stochastic Optimization", ICLR 2015.

source

Merlin.SGD — Type.

SGD

Stochastic Gradient Descent Optimizer.

Arguments

rate: learning rate

[momentum=0.0]: momentum coefficient

[nesterov=false]: use nesterov acceleration or not

source