Functions

Functions

Merlin.Conv

Merlin.Linear

Merlin.Swish

Base.:*

Base.:+

Base.:-

Base.:/

Base.:^

Base.LinAlg.BLAS.gemm

Base.LinAlg.BLAS.gemv

Base.broadcast

Base.broadcast

Base.broadcast

Base.exp

Base.getindex

Base.log

Base.max

Base.mean

Base.reshape

Base.sum

Base.tanh

Base.transpose

Merlin.argmax

Merlin.concat

Merlin.crelu

Merlin.crossentropy

Merlin.dropout

Merlin.elu

Merlin.gemm_batch

Merlin.l2

Merlin.leaky_relu

Merlin.logsoftmax

Merlin.max_batch

Merlin.mse

Merlin.relu

Merlin.selu

Merlin.sigmoid

Merlin.softmax

Merlin.softmax_crossentropy

Activation

Merlin.crelu — Method.

crelu(x::Var)

Concatenated Rectified Linear Unit. The output is twice the size of the input.

\[f(x) = (\max(0,x), \max(0,-x))\]

References

Shang et al., "Understanding and Improving Convolutional Neural Networks via Concatenated Rectified Linear Units", arXiv 2016.

source

Merlin.elu — Method.

elu(x::Var)

Exponential Linear Unit.

References

Clevert et al., "Fast and Accurate Deep Network Learning by Exponential Linear Units (ELUs)", arXiv 2015.

\[f(x) = \begin{cases} x & x > 0 \\ \alpha (e^{x}-1) & x\leq0 \end{cases}\]

where $\alpha=1$.

source

Merlin.leaky_relu — Function.

leaky_relu(x::Var, alpha=0.1)

Leaky Rectified Linear Unit.

\[f(x) = \begin{cases} x & x > 0 \\ \alpha x & x \leq 0 \end{cases}\]

References

Maas et al., "Rectifier Nonlinearities Improve Neural Network Acoustic Models", ICML 2013.

source

Merlin.relu — Method.

relu(x::Var)

Rectified Linear Unit.

\[f(x) = \max(0, x)\]

source

Merlin.selu — Method.

selu(x::Var)

Scaled Exponential Linear Unit.

\[f(x) = \lambda \begin{cases} x & x > 0 \\ \alpha e^{x}-\alpha & x\leq0 \end{cases}\]

where $\lambda=1.0507$ and $\alpha=1.6733$.

References

Klambauer et al., "Self-Normalizing Neural Networks", NIPS 2017.

source

Merlin.sigmoid — Method.

sigmoid(x)

Sigmoid logistic function.

\[f(x) = (1 + \exp(-x))^{-1}\]

source

Merlin.Swish — Type.

Swish

Swish activation function.

\[f(x) = x \cdot \sigma (\beta x)\]

where $\beta$ is a leanable parameter.

References

Ramachandran et al. "Searching for Activation Functions", arXiv 2017.

source

Base.tanh — Method.

tanh(x::Var)

Hyperbolic tangent function.

source

BLAS

Base.LinAlg.BLAS.gemm — Method.

gemm(tA::Char, tB::Char, alpha, A::Var, B::Var)
gemm(A::Var, B::Var, [tA='N'], [tB='N'], [alpha=1])

tA, tB: 'T' (transpose) or 'N' (not transpose)

\[C = \alpha \times \textrm{tA}(A) \times \textrm{tB}(B)\]

T = Float32
A = Var(rand(T,10,5))
B = Var(rand(T,10,7))
C = BLAS.gemm('T', 'N', 1, A, B)

source

Base.LinAlg.BLAS.gemv — Method.

BLAS.gemv(tA::Char, alpha, A::Var, x::Var)

tA: 'T' (transpose) or 'N' (not transpose)

\[y = \alpha \times \textrm{tA}(A) \times x\]

T = Float32
A = Var(rand(T,10,5))
x = Var(rand(T,5))
B = BLAS.gemv('N', 1, A, x)

source

Merlin.gemm_batch — Method.

gemm_batch(tA::Char, tB::Char, alpha, As::Vector{Var}, B::Vector{Var})
gemm_batch(As::Vector{Var}, B::Vector{Var}, [tA='N'], [tB='N'], [alpha=1])

source

Convolution

Merlin.Conv — Type.

Conv(T, filtersize, kwargs...)

W: (W1,W2,...,I,O)

X: (X1,X2,...,I,N)

Y: (Y1,Y2,...,O,N)

where

I: number of input channels

O: number of output channels

N: batch size

T = Float32
conv = Conv(T, (1,1,3,2))
x = CuArray{T}(5,5,3,3)
y = conv(x)

source

Embeddings

Loss

Merlin.crossentropy — Method.

crossentropy(p, q)

Cross-entropy function between p and q.

\[f(x) = -\sum_{x} p(x) \log q(x)\]

p::Var: Var of Vector{Int} or Matrix{Float}. If p is Vector{Int} and p[i] == 0, returns 0.

q::Var: Var of Matrix{Float}

p = Var(rand(0:10,5))
q = softmax(Var(rand(Float32,10,5)))
y = crossentropy(p, q)

source

Merlin.l2 — Method.

l2(x::Var, lambda::Float64)

L2 regularization.

\[y = \frac{\lambda}{2}\left\Vert \mathbf{x} \right\Vert ^{2}\]

x = Var(rand(Float32,10,5))
y = l2(x, 0.01)

source

Merlin.mse — Method.

mse(x1, x2)

Mean Squared Error function between x1 and x2. The mean is calculated over the minibatch. Note that the error is not scaled by 1/2.

source

Merlin.softmax_crossentropy — Method.

softmax_crossentropy(p, x)

Cross-entropy function between p and $softmax(x)$.

\[f(x) = -\sum_{x} p(x) \log q(x)\]

where $q = softmax(x)$

p: Var of Vector{Int} or Matrix{Float}

q: Var of Matrix{Float}

p = Var(rand(0:10,5))
q = Var(rand(Float32,10,5))
y = softmax_crossentropy(p, x)

source

Math

Base.:* — Method.

*(A::Var, B::Var)

source

Base.:+ — Method.

+(x1::Var, x2::Var)
+(a::Number, x::Var)
+(x::Var, a::Number)

source

Base.:- — Method.

-(x1, x2)

source

Base.:/ — Method.

/(x1::Var, a)

source

Base.:^ — Method.

^(x::Var, a::Number)

source

Base.broadcast — Method.

.*(x1::Var, x2::Var)

source

Base.broadcast — Method.

.+(x1::Var, x2::Var)

source

Base.broadcast — Method.

.-(x1::Var, x2::Var)

source

Base.exp — Method.

exp(x)

source

Base.log — Method.

log(x)

source

Base.transpose — Method.

transpose(x)

source

Recurrent

Reduction

Merlin.max_batch — Method.

max_batch(x::Var, dims::Vector{Int})

source

Base.max — Method.

maximum(x::Var, dim::Int)

Returns the maximum value over the given dimension.

x = Var(rand(Float32,10,5))
y = maximum(x, 1)

source

Base.mean — Method.

mean(x, dim::Int)

Computes the average over the given dimension.

source

Base.sum — Method.

sum(x::Var, dim::Int)

Returns the sum over the given dimension.

source

Misc

argmax
concat
dropout
getindex
Linear
lookup
reshape
softmax
logsoftmax
standardize